refactor: split server module, extract puzzle data, drop persistence boilerplate
- server/__init__.py shrinks from 575→41 lines; logic moves to handlers.py,
  protocol.py (WS framing), rooms.py (multiplayer state)
- /world, /observe, /puzzle, /puzzle/attempt now share a single _build_world
  helper; world-rebuild logic is no longer duplicated four times
- puzzle pools (~330 lines of static data) move to puzzles/data.py;
  engine.py shrinks from 481→121 lines
- persistence: init_db runs once via @_with_db decorator instead of on every
  call; chmod runs once at init instead of every connect
