+++ persistence/__init__.py
@@ -1,66 +1,82 @@
 from __future__ import annotations
 
+import functools
 import json
 import sqlite3
 import stat
 from pathlib import Path
-from typing import Any
+from typing import Any, Callable
 
 _DB_PATH = Path.home() / ".nested-worlds" / "worlds.db"
 
+_SCHEMA = """
+    CREATE TABLE IF NOT EXISTS worlds (
+        seed        INTEGER PRIMARY KEY,
+        created_at  TEXT NOT NULL DEFAULT (datetime('now')),
+        node_count  INTEGER,
+        max_depth   INTEGER,
+        min_breadth INTEGER,
+        max_breadth INTEGER
+    );
+
+    CREATE TABLE IF NOT EXISTS agent_runs (
+        id            INTEGER PRIMARY KEY AUTOINCREMENT,
+        agent_name    TEXT NOT NULL,
+        world_seed    INTEGER NOT NULL,
+        started_at    TEXT NOT NULL DEFAULT (datetime('now')),
+        nodes_visited INTEGER,
+        events        TEXT
+    );
+
+    CREATE TABLE IF NOT EXISTS puzzle_results (
+        id          INTEGER PRIMARY KEY AUTOINCREMENT,
+        world_seed  INTEGER NOT NULL,
+        puzzle_name TEXT NOT NULL,
+        result      TEXT NOT NULL,
+        attempts    INTEGER NOT NULL,
+        recorded_at TEXT NOT NULL DEFAULT (datetime('now'))
+    );
+
+    CREATE TABLE IF NOT EXISTS world_mutations (
+        id            INTEGER PRIMARY KEY AUTOINCREMENT,
+        world_seed    INTEGER NOT NULL,
+        node_name     TEXT NOT NULL,
+        mutation_type TEXT NOT NULL,
+        player_name   TEXT,
+        data          TEXT,
+        recorded_at   TEXT NOT NULL DEFAULT (datetime('now'))
+    );
+"""
+
+_initialized: set[Path] = set()
+
 
 def _connect() -> sqlite3.Connection:
-    _DB_PATH.parent.mkdir(parents=True, exist_ok=True)
     conn = sqlite3.connect(_DB_PATH)
     conn.execute("PRAGMA journal_mode=WAL")
-    _DB_PATH.chmod(stat.S_IRUSR | stat.S_IWUSR)  # 0o600 — owner read/write only
     return conn
 
 
 def init_db() -> None:
+    _DB_PATH.parent.mkdir(parents=True, exist_ok=True)
     with _connect() as conn:
-        conn.executescript("""
-            CREATE TABLE IF NOT EXISTS worlds (
-                seed        INTEGER PRIMARY KEY,
-                created_at  TEXT NOT NULL DEFAULT (datetime('now')),
-                node_count  INTEGER,
-                max_depth   INTEGER,
-                min_breadth INTEGER,
-                max_breadth INTEGER
-            );
-
-            CREATE TABLE IF NOT EXISTS agent_runs (
-                id            INTEGER PRIMARY KEY AUTOINCREMENT,
-                agent_name    TEXT NOT NULL,
-                world_seed    INTEGER NOT NULL,
-                started_at    TEXT NOT NULL DEFAULT (datetime('now')),
-                nodes_visited INTEGER,
-                events        TEXT
-            );
-
-            CREATE TABLE IF NOT EXISTS puzzle_results (
-                id          INTEGER PRIMARY KEY AUTOINCREMENT,
-                world_seed  INTEGER NOT NULL,
-                puzzle_name TEXT NOT NULL,
-                result      TEXT NOT NULL,
-                attempts    INTEGER NOT NULL,
-                recorded_at TEXT NOT NULL DEFAULT (datetime('now'))
-            );
-
-            CREATE TABLE IF NOT EXISTS world_mutations (
-                id            INTEGER PRIMARY KEY AUTOINCREMENT,
-                world_seed    INTEGER NOT NULL,
-                node_name     TEXT NOT NULL,
-                mutation_type TEXT NOT NULL,
-                player_name   TEXT,
-                data          TEXT,
-                recorded_at   TEXT NOT NULL DEFAULT (datetime('now'))
-            );
-        """)
+        conn.executescript(_SCHEMA)
+    # 0o600 — owner read/write only.  Set once on init, not per-connect.
+    _DB_PATH.chmod(stat.S_IRUSR | stat.S_IWUSR)
+    _initialized.add(_DB_PATH)
 
 
+def _with_db(fn: Callable) -> Callable:
+    @functools.wraps(fn)
+    def wrapper(*args, **kwargs):
+        if _DB_PATH not in _initialized:
+            init_db()
+        return fn(*args, **kwargs)
+    return wrapper
+
+
+@_with_db
 def save_world(seed: int, node_count: int, max_depth: int, min_breadth: int, max_breadth: int) -> None:
-    init_db()
     with _connect() as conn:
         conn.execute(
             """INSERT OR REPLACE INTO worlds (seed, node_count, max_depth, min_breadth, max_breadth)
@@ -69,8 +85,8 @@ def save_world(seed: int, node_count: int, max_depth: int, min_breadth: int, max
         )
 
 
+@_with_db
 def save_agent_run(agent_name: str, world_seed: int, nodes_visited: int, events: list[dict[str, Any]]) -> int:
-    init_db()
     with _connect() as conn:
         cur = conn.execute(
             """INSERT INTO agent_runs (agent_name, world_seed, nodes_visited, events)
@@ -80,8 +96,8 @@ def save_agent_run(agent_name: str, world_seed: int, nodes_visited: int, events:
         return cur.lastrowid
 
 
+@_with_db
 def save_puzzle_result(world_seed: int, puzzle_name: str, result: str, attempts: int) -> None:
-    init_db()
     with _connect() as conn:
         conn.execute(
             """INSERT INTO puzzle_results (world_seed, puzzle_name, result, attempts)
@@ -90,8 +106,8 @@ def save_puzzle_result(world_seed: int, puzzle_name: str, result: str, attempts:
         )
 
 
+@_with_db
 def list_worlds() -> list[dict[str, Any]]:
-    init_db()
     with _connect() as conn:
         rows = conn.execute(
             "SELECT seed, created_at, node_count, max_depth FROM worlds ORDER BY created_at DESC"
@@ -99,9 +115,9 @@ def list_worlds() -> list[dict[str, Any]]:
         return [{"seed": r[0], "created_at": r[1], "node_count": r[2], "max_depth": r[3]} for r in rows]
 
 
+@_with_db
 def record_mutation(world_seed: int, node_name: str, mutation_type: str,
                     player_name: str | None, data: dict) -> None:
-    init_db()
     with _connect() as conn:
         conn.execute(
             """INSERT INTO world_mutations (world_seed, node_name, mutation_type, player_name, data)
@@ -110,8 +126,8 @@ def record_mutation(world_seed: int, node_name: str, mutation_type: str,
         )
 
 
+@_with_db
 def get_mutations(world_seed: int, limit: int = 50) -> list[dict[str, Any]]:
-    init_db()
     with _connect() as conn:
         rows = conn.execute(
             """SELECT node_name, mutation_type, player_name, data, recorded_at
@@ -124,8 +140,8 @@ def get_mutations(world_seed: int, limit: int = 50) -> list[dict[str, Any]]:
                 for r in rows]
 
 
+@_with_db
 def get_agent_runs(world_seed: int) -> list[dict[str, Any]]:
-    init_db()
     with _connect() as conn:
         rows = conn.execute(
             """SELECT agent_name, started_at, nodes_visited
